Refactor unit test http helper extensions
